+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 -   repo: https://github.com/dannyvassallo/cloudformation-guard
-    rev: pre-commit-v0.0.15
+    rev: pre-commit-v0.0.16
     hooks:
     -   id: cfn-guard
         args: [validate]
